fix: The program now takes non-contiguous partitions into account
Before this commit, the program wouldn't take into account the fact that
partitions can be non-contiguous, meaning that the start of a partition
can be far apart from the end of the previous one, thus creating
specific "free spaces" in the disk.

The program now also allows the user to reset the disk selection and
partition layout.